refactor: move WASM state to dedicated module
